# Fix caching of small inline images in `Parser.makeInlineImage` (issue 8790)