--- a/package.json
+++ b/package.json
@@ -27,6 +27,7 @@
     "streamqueue": "^1.1.1",
     "systemjs": "^0.20.7",
     "systemjs-plugin-babel": "0.0.21",
+    "ttest": "^1.1.0",
     "typogr": "^0.6.6",
     "uglify-js": "^2.6.1",
     "vinyl-fs": "^2.4.4",